+++ ometha/__init__.py
@@ -1,0 +1,3 @@
+from ._version import __version__ as _version
+
+__version__ = _version
