+++ ometha/helpers.py
@@ -1,6 +1,11 @@
+import re
+import sys
 import time
 
-from colorama import Fore, Style, init
+from colorama import Fore, Style
+from loguru import logger
+from lxml import etree
+from requests.exceptions import ConnectionError, HTTPError, RetryError, Timeout
 
 # define global variables reused throughout the code
 SEP_LINE = "--------------------------------------\n"
@@ -11,8 +16,7 @@ TIMESTR = time.strftime("%Y-%m-%dT%H:%M:%SZ")
 NAMESPACE = "{http://www.openarchives.org/OAI/2.0/}"
 ISODATEREGEX = "(?:19|20)[0-9]{2}-(?:(?:0[1-9]|1[0-2])-(?:0[1-9]|1[0-9]|2[0-9])|(?:(?!02)(?:0[1-9]|1[0-2])-(?:30))|(?:(?:0[13578]|1[02])-31))"
 URLREGEX = r"https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)"
-# TODO read the version from setup.py
-__version__ = "2.0.0"
+
 # initialize all parameters in a dict shortened as PRM
 PRM = {
     "b_url": None,  # base url: str
@@ -33,3 +37,55 @@ PRM = {
     "mode": None,  # mode: str "ui" or "cli"
     "exp_type": None,  # export type either "xml" or "json"
 }
+
+
+def print_and_log(message, logger, type: str, end="\n"):
+    print(message, end)
+    for placeholder in [SEP_LINE, ACHTUNG, INFO, FEHLER]:
+        if placeholder in message:
+            message = message.replace(placeholder, "")
+    if type == "info":
+        logger.info(message)
+    elif type == "warning":
+        logger.warning(message)
+
+
+def handle_error(e, mode, url=None):
+    error_messages = {
+        Timeout: "Timeout determining list size. The API is not reachable.",
+        RetryError: "Identifier harvesting aborted due to too many retries. Is the API reachable?",
+        HTTPError: "Identifier harvesting aborted due to an HTTP error.",
+    }
+    if type(e) is ConnectionError:
+        if "404" in str(e):
+            log_critical_and_print_and_exit(
+                "The API is not reachable. Is the URL correct?", mode
+            )
+        elif errors := re.findall(r"error\scode=['\"](.+)['\"]>(.*)<\\error", str(e)):
+            log_critical_and_print_and_exit(
+                f"{FEHLER} API error: {errors[0][0]}/{errors[0][1]} at {url}", mode
+            )
+    elif type(e) in error_messages:
+        log_critical_and_print_and_exit(error_messages[type(e)], mode, e)
+    else:
+        log_critical_and_print_and_exit("An unexpected error occurred.", mode, e)
+
+
+def log_critical_and_print_and_exit(message, mode=None, exception=None):
+    # FIXME this leads to ugly log-files because the separator is sometimes printed
+    logger.critical(message)
+    if exception:
+        logger.exception("Exception details:", exc_info=exception)
+    if mode == "ui" and input(f"{message}\nDrücken Sie Enter zum Beenden..."):
+        sys.exit()
+
+
+def isinvalid_xml_content(response, url, mode):
+    try:
+        root = etree.XML(response.content)
+    except etree.XMLSyntaxError as e:
+        log_critical_and_print_and_exit(
+            f"XML Syntax Error in the API response, probably no valid XML ({url}): '{e}'",
+            mode,
+        )
+    return root
